refactor(common): promote Symbol newtype to shared common module
Move Symbol and Symbols from market_data::types into the shared common
module so that sibling endpoint modules (fundamentals, etc.) can reuse
the ticker-symbol newtype without depending on each other.
